Refactor: 오늘의 추천 문제 API를 Study에서 분리
- assignments 폴더에 해당 API 정의
