+++ core/urls.py
@@ -4,4 +4,5 @@ urlpatterns = [
     path("", include("core.auth.urls")),
     path("", include("core.user.urls")),
     path("", include("core.study.urls")),
+    path("", include("core.assignments.urls")),
 ]
